Add files via upload
This requires two csv files storing the data for the players ELO, and another for the players ELO for the last calendar year (YELO). At the moment there is a website where these are stored, but it is not updated regularly. It then calculates the odds of two selected players winning. You can then enter the odds given by a sports betting website and compare your odds to those to see if your prediction is different. 

Ultimately, I would want to automate the ELO documents so they are updated regularly and then compare over the long run if I have a more accurate model than a sports betting website by betting against them (theoretically) and seeing if I can "win".